- Implemented `TrimedGNM` class for Hi-C to use. - saveModel() and loadModel() are also adjusted to the new class.
