+++ prody/dynamics/gnm.py
@@ -16,7 +16,7 @@ from prody.utilities import importLA, checkCoords
 from .nma import NMA
 from .gamma import Gamma
 
-__all__ = ['GNM', 'calcGNM']
+__all__ = ['GNM', 'calcGNM', 'TrimedGNM']
 
 ZERO = 1e-6
 
@@ -429,3 +429,52 @@ def calcGNM(pdb, selstr='calpha', cutoff=15., gamma=1., n_modes=20,
     gnm.buildKirchhoff(sel, cutoff, gamma)
     gnm.calcModes(n_modes, zeros, hinges=hinges)
     return gnm, sel
+
+class TrimedGNM(GNM):
+    def __init__(self, name='Unknown', mask=False, useTrimed=True):
+        super(TrimedGNM, self).__init__(name)
+        self.mask = False
+        self.useTrimed = useTrimed
+
+        if not np.isscalar(mask):
+            self.mask = np.array(mask)
+
+    def numAtoms(self):
+        """Returns number of atoms."""
+
+        if self.useTrimed or np.isscalar(self.mask):
+            return self._n_atoms
+        else:
+            return len(self.mask)
+
+    def getArray(self):
+        """Returns a copy of eigenvectors array."""
+
+        if self._array is None: return None
+
+        array = self._array.copy()
+
+        if self.useTrimed or np.isscalar(self.mask):
+            return array
+
+        mask = ~self.mask.copy()
+        N = len(mask)
+        n, m = array.shape
+        whole_array = np.zeros((N,m))
+        mask = np.expand_dims(mask, axis=1)
+        mask = mask.repeat(m, axis=1)
+        whole_array[mask] = array.flatten()
+        return whole_array
+
+    getEigvecs = getArray
+
+    def _getArray(self):
+        """Returns eigenvectors array. The function returns 
+        a copy of the array if useTrimed is ``True``."""
+
+        if self._array is None: return None
+
+        if self.useTrimed or np.isscalar(self.mask):
+            return self._array
+        else:
+            return self.getArray()
